\name{true_label}
\alias{true_label}
\docType{data}
\title{
True clustering lables of 200 simulated cases
}
\description{
Dataset used to demonstrate the use of WSNF. 
}
\usage{
data(true_label)
}
\format{
A vector indicating the true clustering lables of 200 cases\cr
Refer to mirPLS: a partial linear structure identifier method for cancer subtyping using MicroRNAs. Peifeng Ruan, Shuang Wang, Hua Liang.

}
\etrue_labelamples{
data(true_label)
}
\keyword{datasets}
