\name{x}
\alias{x}
\docType{data}
\title{
A sample of simulated miRNA data
}
\description{
Dataset used to demonstrate the use of WSNF. 
}
\usage{
data(x)
}
\format{
A 300 subjects * 300 miRNAs matrix. The first 6 columns are 6 miRNAs have non-linear associations, the next 14 columns are 14 miRNAs have non-linear associations and the others are miRNAs have no associations.\cr
Refer to simulation part of mirPLS: a partial linear structure identifier algorithm for miRNA data. Peifeng Ruan, Shuang Wang, Hua Liang.

}
\examples{
data(x)
}
\keyword{datasets}
