\name{y}
\alias{y}
\docType{data}
\title{
A sample of simulated outcome data
}
\description{
Dataset used to demonstrate the use of WSNF. 
}
\usage{
data(y)
}
\format{
A vector indicating whether a subject is a case or a control. 1s represents cases and 0s represents controls.\cr
Refer to mirPLS: a partial linear structure identifier method for cancer subtyping using MicroRNAs. Peifeng Ruan, Shuang Wang, Hua Liang.

}
\examples{
data(y)
}
\keyword{datasets}
